modify icon finish anti Spam v1 Complete bad words list

+++ Events/Moderation/messageAntiRaid.js
@@ -2,8 +2,15 @@ const { Message, Client, PermissionFlagsBits } = require("discord.js");
 
 let badWordsArray = [
     "Fuck",
+    "fuck",
     "gueule",
-    "Pute"
+    "Pute",
+    "pute",
+    "Connard",
+    "connard",
+    "FDP",
+    "fdp",
+    "sale chien"
 ]
 
 module.exports = {
@@ -40,12 +47,16 @@ module.exports = {
 
             badWordsArray.forEach(badWord => {
                 if(message.content.includes(`${badWord}`)) {
-                    message.delete()
-                    return message.channel.send({
-                        content: `${message.author}, you'are not allowed to post this type of message !!`
-                    }).then((msg) => {
-                        setTimeout(() => msg.delete(), 10000)
-                    })
+                    if(message.member.roles.cache.has("1137411181485768824") || message.member.permissions.has(PermissionFlagsBits.ModerateMembers)) {
+                        return;
+                    } else {
+                        message.delete()
+                        return message.channel.send({
+                            content: `${message.author}, you'are not allowed to post this type of message !!`
+                        }).then((msg) => {
+                            setTimeout(() => msg.delete(), 10000)
+                        })
+                    }
                 }
             });
